+++ rustls/src/msgs/cookie.rs
@@ -1,3 +1,5 @@
+use crate::msgs::codec::{Codec, Reader};
+
 #[derive(Copy)]
 pub struct Cookie {
     len: usize,
